修复DemandPredictionServiceImpl中的返回值类型，将courierMapper.selectCount(wrapper)的返回值转换为int，以确保返回类型一致性。

+++ backend/src/main/java/com/ruralexpress/service/impl/DemandPredictionServiceImpl.java
@@ -320,7 +320,7 @@ public class DemandPredictionServiceImpl implements DemandPredictionService {
         wrapper.eq(Courier::getStatus, 1) // 启用状态
                .eq(Courier::getServiceArea, area);
         
-        return courierMapper.selectCount(wrapper);
+        return courierMapper.selectCount(wrapper).intValue();
     }
     
     /**
